Moves route related code from occlient.go to a new file and adds more tests
