+++ pkg/occlient/occlient_test.go
@@ -17,7 +17,6 @@ import (
 	dockerapi "github.com/openshift/api/image/docker10"
 	dockerapiv10 "github.com/openshift/api/image/docker10"
 	imagev1 "github.com/openshift/api/image/v1"
-	routev1 "github.com/openshift/api/route/v1"
 	applabels "github.com/openshift/odo/pkg/application/labels"
 	componentlabels "github.com/openshift/odo/pkg/component/labels"
 	"github.com/openshift/odo/pkg/config"
@@ -309,142 +308,6 @@ func fakePlanInstanceCreateParameterSchemasRaw() ([][]byte, error) {
 	data = append(data, planInstanceCreateParameterSchemaRaw2)
 
 	return data, nil
-}
-
-func TestCreateRoute(t *testing.T) {
-	tests := []struct {
-		name       string
-		urlName    string
-		service    string
-		portNumber intstr.IntOrString
-		labels     map[string]string
-		wantErr    bool
-		existingDC appsv1.DeploymentConfig
-		secureURL  bool
-		path       string
-	}{
-		{
-			name:       "Case : mailserver",
-			urlName:    "mailserver",
-			service:    "mailserver",
-			portNumber: intstr.FromInt(8080),
-			labels: map[string]string{
-				"SLA":                        "High",
-				"app.kubernetes.io/instance": "backend",
-				"app.kubernetes.io/name":     "python",
-			},
-			wantErr:    false,
-			existingDC: *fakeDeploymentConfig("mailserver", "", nil, nil, t),
-		},
-
-		{
-			name:       "Case : blog (urlName is different than service)",
-			urlName:    "example",
-			service:    "blog",
-			portNumber: intstr.FromInt(9100),
-			labels: map[string]string{
-				"SLA":                        "High",
-				"app.kubernetes.io/instance": "backend",
-				"app.kubernetes.io/name":     "golang",
-			},
-			wantErr:    false,
-			existingDC: *fakeDeploymentConfig("blog", "", nil, nil, t),
-		},
-
-		{
-			name:       "Case : secure url",
-			urlName:    "example",
-			service:    "blog",
-			portNumber: intstr.FromInt(9100),
-			labels: map[string]string{
-				"SLA":                        "High",
-				"app.kubernetes.io/instance": "backend",
-				"app.kubernetes.io/name":     "golang",
-			},
-			wantErr:    false,
-			existingDC: *fakeDeploymentConfig("blog", "", nil, nil, t),
-			secureURL:  true,
-		},
-
-		{
-			name:       "Case : specify a path",
-			urlName:    "example",
-			service:    "blog",
-			portNumber: intstr.FromInt(9100),
-			labels: map[string]string{
-				"SLA":                        "High",
-				"app.kubernetes.io/instance": "backend",
-				"app.kubernetes.io/name":     "golang",
-			},
-			wantErr:    false,
-			existingDC: *fakeDeploymentConfig("blog", "", nil, nil, t),
-			path:       "/testpath",
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			// initialising the fakeclient
-			fkclient, fkclientset := FakeNew()
-
-			ownerReferences := GenerateOwnerReference(&tt.existingDC)
-
-			fkclientset.AppsClientset.PrependReactor("get", "deploymentconfigs", func(action ktesting.Action) (bool, runtime.Object, error) {
-				return true, &tt.existingDC, nil
-			})
-
-			createdRoute, err := fkclient.CreateRoute(tt.urlName, tt.service, tt.portNumber, tt.labels, tt.secureURL, tt.path, ownerReferences)
-
-			if tt.secureURL {
-				wantedTLSConfig := &routev1.TLSConfig{
-					Termination:                   routev1.TLSTerminationEdge,
-					InsecureEdgeTerminationPolicy: routev1.InsecureEdgeTerminationPolicyRedirect,
-				}
-				if !reflect.DeepEqual(createdRoute.Spec.TLS, wantedTLSConfig) {
-					t.Errorf("tls config is different, wanted %v, got %v", wantedTLSConfig, createdRoute.Spec.TLS)
-				}
-			} else {
-				if createdRoute.Spec.TLS != nil {
-					t.Errorf("tls config is set for a non secure url")
-				}
-				if tt.path == "" && createdRoute.Spec.Path != "/" {
-					t.Errorf("expect path: /, but got path: %v", createdRoute.Spec.Path)
-				} else if tt.path != "" && createdRoute.Spec.Path != tt.path {
-					t.Errorf("expect path: %v, but got path: %v", tt.path, createdRoute.Spec.Path)
-				}
-			}
-
-			// Checks for error in positive cases
-			if !tt.wantErr == (err != nil) {
-				t.Errorf(" client.CreateRoute(string, labels) unexpected error %v, wantErr %v", err, tt.wantErr)
-			}
-
-			// Check for validating actions performed
-			if len(fkclientset.RouteClientset.Actions()) != 1 {
-				t.Errorf("expected 1 action in CreateRoute got: %v", fkclientset.RouteClientset.Actions())
-			}
-			// Checks for return values in positive cases
-			if err == nil {
-				createdRoute := fkclientset.RouteClientset.Actions()[0].(ktesting.CreateAction).GetObject().(*routev1.Route)
-				// created route should be labeled with labels passed to CreateRoute
-				if !reflect.DeepEqual(createdRoute.Labels, tt.labels) {
-					t.Errorf("labels in created route is not matching expected labels, expected: %v, got: %v", tt.labels, createdRoute.Labels)
-				}
-				// route name and service that route is pointg to should match
-				if createdRoute.Spec.To.Name != tt.service {
-					t.Errorf("route is not matching to expected service name, expected: %s, got %s", tt.service, createdRoute)
-				}
-				if createdRoute.Name != tt.urlName {
-					t.Errorf("route name is not matching to expected route name, expected: %s, got %s", tt.urlName, createdRoute.Name)
-				}
-				if createdRoute.Spec.To.Name != tt.service {
-					t.Errorf("service name is not matching to expected service name, expected: %s, got %s", tt.service, createdRoute.Spec.To.Name)
-				}
-				if createdRoute.Spec.Port.TargetPort != tt.portNumber {
-					t.Errorf("port number is not matching to expected port number, expected: %v, got %v", tt.portNumber, createdRoute.Spec.Port.TargetPort)
-				}
-			}
-		})
-	}
 }
 
 func TestAddLabelsToArgs(t *testing.T) {
